requirements.txt
pytest
playwright
openai
ollama
requests
pandas
openpyxl
pyyaml
python-dotenv
pytest-html
allure-pytest